requirements.txt
langchain==0.0.135
pinecone-client==2.2.1
python-dotenv==1.0.0
tiktoken==0.3.3
